+++ packages/engine-core/src/index.ts
@@ -6,3 +6,4 @@ export * from "./types.js";
 export * from "./typographySelectors.js";
 export * from "./typographyComposition.js";
 export * from "./colorHarmony.js";
+export * from "./typographyMotionGrammar.js";